Move FundingLink to top level type
