+++ Sources/App/Core/Github.swift
@@ -249,53 +249,6 @@ extension Github {
         var htmlUrl: String
     }
 
-    struct FundingLink: Codable, Equatable {
-
-        enum Platform: String, Codable {
-            case communityBridge
-            case customUrl
-            case gitHub
-            case issueHunt
-            case koFi
-            case lfxCrowdfunding
-            case liberaPay
-            case openCollective
-            case otechie
-            case patreon
-            case tideLift
-
-            private static let gitHubApiEncodings: [String: Platform] = [
-                "COMMUNITY_BRIDGE": .communityBridge,
-                "CUSTOM": .customUrl,
-                "GITHUB": .gitHub,
-                "ISSUEHUNT": .issueHunt,
-                "KO_FI": .koFi,
-                "LFX_CROWDFUNDING": .lfxCrowdfunding,
-                "LIBERAPAY": .liberaPay,
-                "OPEN_COLLECTIVE": .openCollective,
-                "OTECHIE": .otechie,
-                "PATREON": .patreon,
-                "TIDELIFT": .tideLift
-            ]
-
-            init(from decoder: Decoder) throws {
-                let container = try decoder.singleValueContainer()
-                let stringValue = try container.decode(String.self)
-
-                if let value = Platform(rawValue: stringValue) {
-                    self = value
-                } else if let value = Platform.gitHubApiEncodings[stringValue] {
-                    self = value
-                } else {
-                    throw DecodingError.dataCorruptedError(in: container, debugDescription: "Unable to decode \(stringValue)")
-                }
-            }
-        }
-
-        var platform: Platform
-        var url: String
-    }
-
     struct Metadata: Decodable, Equatable {
         static func query(owner: String, repository: String) -> GraphQLQuery {
             // Go to https://developer.github.com/v4/explorer/ to run query manually
